refactor(memfs): keep sync reminders in reminders module
Keep low-level MemFS git state in memory-git while leaving post-turn reminder formatting and delivery orchestration under reminders.
